Refactor: HealthSystem tambien maneja los comandos (Comandos deberian ir aparte), se agregan comentarios para facilitar comprension
